Finally get __INITIAL_STATE__ & hydration working
Vue's documentation is infuriating because it's classic JavaScript-world
shit. It took me most of the day to debug a stupid problem because the
API docs don't list argument and result types of the `createRenderer`
and `createBundleRenderer` functions. I had to hunt through the Vue HackerNews example
code line by line, tracing function calls to figure out what to pass to
`createRenderer` and `renderToString` in order to make it work. It's pathetic.

Anyway, it's properly working now, with much cruft removed.
Architecturally it's actually pretty sound; this commit represents a
good base from which to work going forward.